move independent types away from core_types and resolve circular import problems
