+++ lukefi/metsi/app/export_handlers/rm_timber.py
@@ -3,7 +3,7 @@ from collections import defaultdict
 from lukefi.metsi.app.app_types import SimResults
 from lukefi.metsi.app.file_io import row_writer
 from lukefi.metsi.domain.collected_types import CrossCutResult
-from lukefi.metsi.sim.core_types import CollectedData
+from lukefi.metsi.sim.collected_data import CollectedData
 
 
 def scan_operation_type_for_event(year: int, cross_cut: list[CrossCutResult]) -> str:
