+++ lukefi/metsi/app/file_io.py
@@ -19,9 +19,9 @@ from lukefi.metsi.app.app_types import ExportableContainer
 from lukefi.metsi.app.app_types import SimResults, ForestOpPayload
 from lukefi.metsi.data.layered_model import PossiblyLayered
 from lukefi.metsi.domain.forestry_types import StandList, ForestStand
-from lukefi.metsi.sim.core_types import CollectedData
 from lukefi.metsi.data.formats.declarative_conversion import Conversion
 from lukefi.metsi.app.utils import MetsiException
+from lukefi.metsi.sim.collected_data import CollectedData
 
 StandReader = Callable[[str | Path], StandList]
 StandWriter = Callable[[Path, ExportableContainer[PossiblyLayered[ForestStand]]], None]
