+++ lukefi/metsi/domain/data_collection/net_present_value.py
@@ -1,4 +1,4 @@
-from lukefi.metsi.sim.core_types import CollectedData, OpTuple
+from lukefi.metsi.sim.collected_data import CollectedData, OpTuple
 from lukefi.metsi.data.model import ForestStand
 from lukefi.metsi.domain.collected_types import CrossCutResult, NPVResult, PriceableOperationInfo
 from lukefi.metsi.domain.utils.file_io import get_renewal_costs_as_dict, get_land_values_as_dict
